Configures project for Docker and venv
Sets up a Dockerfile for containerization and configures the project to use a virtual environment (venv).

This change standardizes the development environment by using a Docker container and a virtual environment for dependency management. It also adds necessary dependencies like Jupyter and shap in requirements.txt and sets the project to use ds_venv.